Apply the Apache 2.0 license to every file
I have finally chosen the Apache 2.0 license over MIT or ISC.
But should this notice be included at the top of every file?
In doubt, I prefer to do it that way. I will also have to see if a
NOTICE file is needed.

Please note that this license doesn't apply to previous commits.