Abstract the API
Everything in the API code that was too HTTP-specific was removed
and put in the drivers/http-json.js file (in comments for now).
The API still works with URL routing and still returns HTTP codes
however; but like that it is not tied to HTTP specifically and can
be called directly in the code with a utility like "lucca-caller"
that I will soon publish on GitHub.

I now need to write some documentation on the abstraction
(especially for the format of the request and response objects)
and the HTTP-JSON driver that will make the API pluggable again
into the Koa server.